Antigravity/requirements.txt
flask
flask-bcrypt
flask-sqlalchemy
pyjwt
cryptography
